+++ ConsoleUI/ConsoleUI.csproj
@@ -9,7 +9,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Core\Core.csproj" />
-    <ProjectReference Include="..\Library\Library.csproj" />
+    <ProjectReference Include="..\Library\Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
