--- Core/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace Core
-{
-    public class Class1
-    {
-
-    }
-}
