+++ Core/Models/ExcercisesAPI.cs
