+++ Core/Models/Schedule.cs
