+++ Core/Service/Interfaces/IWorkoutService.cs
@@ -1,0 +1,8 @@
+﻿
+namespace WorkoutApp.ApplicationLogic.Interfaces
+{
+    public interface IWorkoutService
+    {
+        public void AddWorkout(string name);
+    }
+}
