--- Library/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace Library
-{
-    public class Class1
-    {
-
-    }
-}
