+++ Library/Data/ApplicationDbContext.cs
@@ -1,13 +1,54 @@
-﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
+﻿using System;
 
-namespace Library.Data
+using System.Reflection.Metadata;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Library.Models;
+
+
+namespace Library.Data;
+
+public class ApplicationDbContext : IdentityDbContext
 {
-    public class ApplicationDbContext : IdentityDbContext
+    private readonly Action<ApplicationDbContext, ModelBuilder> _modelCustomizer;
+
+    #region Constructors
+    public ApplicationDbContext()
     {
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
-            : base(options)
+
+    }
+
+    public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options, Action<ApplicationDbContext, ModelBuilder> modelCustomizer = null)
+        : base(options)
+    {
+        _modelCustomizer = modelCustomizer;
+    }
+    #endregion
+
+    public DbSet<Workout> Workouts => Set<Workout>();
+    public DbSet<ExercisesAPI> ExercisesAPIs { get; set; }
+    public DbSet<Schedule> Schedules { get; set; }
+
+    public DbSet<Nutrition> Nutritions { get; set; }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        if (!optionsBuilder.IsConfigured)
         {
+            optionsBuilder.UseSqlServer(
+                @"Server=tcp:workout.database.windows.net,1433;InitialCatalog=WorkoutApp;PersistSecurityInfo=False;UserID=WorkoutAppName;Password=Passwordisverysafe!;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;ConnectionTimeout=30;");
+        }
+    }
+
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        //modelBuilder.Entity<UrlResource>().HasNoKey()
+        //    .ToView("AllResources");
+
+        if (_modelCustomizer is not null)
+        {
+            _modelCustomizer(this, modelBuilder);
         }
     }
 }
