--- Library/Service/Interfaces/IWorkoutService.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Library.Models;
-
-namespace WorkoutApp.ApplicationLogic.Interfaces
-{
-    public interface IWorkoutService
-    {
-        public static void AddWorkout(Workout workout)
-        {
-
-        }
-    }
-}
