+++ Prototype.sln
@@ -15,9 +15,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain+Application", "Domai
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "Core\Core.csproj", "{21BACC8D-1D2F-40CF-9096-B91BF3FBA3C1}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Library", "Library\Library.csproj", "{5A0DB5BF-2A28-492C-A045-3F5B22367240}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Infrastructure", "Library\Infrastructure.csproj", "{5A0DB5BF-2A28-492C-A045-3F5B22367240}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleUI", "ConsoleUI\ConsoleUI.csproj", "{570DF82A-B69D-4A03-A717-C027B2A0CF06}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{475898BB-69CE-40DE-A784-27B79DCD336B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -50,7 +52,7 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{AFAF1E8D-94C1-4C47-AA58-F4A9AA1A697D} = {9FEA228B-8959-409C-95F7-FB08B91C86A9}
+		{AFAF1E8D-94C1-4C47-AA58-F4A9AA1A697D} = {475898BB-69CE-40DE-A784-27B79DCD336B}
 		{3C95A167-D0DB-4960-8E5E-2F8589F5D488} = {B3092A51-1DF6-4AEB-99BA-51E053F8A86A}
 		{21BACC8D-1D2F-40CF-9096-B91BF3FBA3C1} = {192D8B36-6A61-4A6F-A015-8F051E380621}
 		{5A0DB5BF-2A28-492C-A045-3F5B22367240} = {9FEA228B-8959-409C-95F7-FB08B91C86A9}
