+++ Web/Pages/_ViewImports.cshtml
@@ -1,7 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
 @using Web
 @using Web.Data
-@using Library.Data;
-@using Library.Service;
 @namespace Web.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
