+++ Web/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Server=tcp:workout.database.windows.net,1433;Initial Catalog=WorkoutApp;Persist Security Info=False;User ID=WorkoutAppName;Password=Passwordisverysafe!;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+    "DefaultConnection": "Server=tcp:workout.database.windows.net,1433;PersistSecurityInfo=False;UserID=WorkoutAppName;Password=Passwordisverysafe!;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;ConnectionTimeout=30;"
   },
   "Logging": {
     "LogLevel": {
