+++ WorkoutApp.Tests/WorkoutApp.Tests.csproj
@@ -29,7 +29,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Library\Library.csproj" />
+    <ProjectReference Include="..\Library\Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
